requirements.txt
-e git+https://github.com/xiilei/dwml.git@32c5b4fc963c160a0af655c1791f447179f8e77e#egg=dwml
lxml==5.2.2
more-itertools==10.3.0
pillow==10.4.0
python-docx==1.1.2
setuptools==69.5.1
typing_extensions==4.12.2
webcolors==24.6.0
wheel==0.43.0
